src/drones/ugv/app/requirements.txt
paho-mqtt>=1.6.1
pyserial>=3.5
opencv-python-headless>=4.8.0
numpy>=1.24.0
